pyproject.toml
[project]
name = "localpdfembed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    "pymupdf4llm>=0.0.27",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
]
